backend/ml/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
openai-whisper>=20231117
openai>=1.0.0
python-toon>=0.0.4
moviepy>=1.0.3
imageio-ffmpeg>=0.4.8
opencv-python-headless>=4.8.0
scenedetect[opencv]>=0.6.2
scikit-learn>=1.3.0
soundfile>=0.12.0
pytesseract>=0.3.10
face-recognition>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.66.0
python-dotenv>=1.0.0